// src/app/models/conductor.ts
export class Conductor {

    id: number;
	nombre: string;
	apellidoPaterno: string;
	apellidoMaterno: string;
	tipoDocumento: string;
	nroDocumento: string;
	telefono: string;
	correo: string;
	direccion: string;
	sexo: string;
	estadoCivil: string;
	fechaNacimiento: string;
	observacion: string;
	estado: string;
	nroLicenciaCorrelativo: string;
	claseCategoria: string;
	estadoLicencia: string;
	fechaExpedicion: string;
	fechaRevalidacion: string;
	restricciones: string;
}